Update handling of child product for product ranges
This commit addresses issue #1253.

Currently, the offers app doesn't really consider child products;
behaviour is undefined. This commit addresses the fact: for now, we
assume that child and parent products are tightly coupled, and forbid
child products in a range. You can only add the parent product (and
with it all children).

This requirement might be lifted in the future, but that will mean
additional complexity and we will see if it's necessary.

Closes #1253.