Add validation rules for options on child products
As child products options are handled on the parents product page, we
can't allow per-product options on child products.

Personally, I think per-product options are confusing anyway; IMHO it
should only be handled by the product class, as is done for product
attributes. But that will be addressed after checking with
@codeinthehole.