Correctly handle invalid parent products
The ProductCreateUpdateView had untested logic to handle invalid parent
products. Unfortunately, it didn't work; the check returned a redirect,
which later was tried to access like a product.

The only place to issue a redirect for invalid products is the dispatch
method, so an altered check has been moved there.

This time with tests...