Fix bug with country field in shipping address form
Before this change, if you submitted a shipping address then navigated
back to the shipping address page, the selected country was lost. It's
possible this has been an issue for a while as the sandbox used to
default to only one shipping country which meant the country field was
removed.

The fix was to convert the country_id into a Country instance when
building the initial data for the form.