Use template egg loader when running the test suite
We wrongly had django-tables2 in the requirements.txt; it was removed
from 4af10d9. This means it was now only installed as an egg. As the
test suite relies on a template from tables2, the test suite
consequently was failing.
Adding the egg loader to the test suite config fixes the problem.