Change domain of call-home pixel
Since Tangent are no longer sponsoring Oscar (at least not to the extend
as previously), it makes sense for the call-home pixel to be controlled
by the new Oscar organisation rather than Tangent. Hence we change the
domain to point to a new server.

Note, the analytics.oscarcommerce.com domain doesn't resolve right now
as the oscarcommerce.com domain is being transferred and I can't add new
records. This should be resolved within the next week or so.

Fixes #1547