Use a regular settings file for tests
I'm not sure why @codeinthehole, back in the days,  opted to
manually configure settings instead of using a regular settings file for
the test suite. Most likely the reasons for doing so have disappeared.

So I turned our tests/config.py into a normal settings file, and use
that accordingly. That means we do things the standard Django way, and
it's easier to drop into a session with the test settings.