Delay setting offer condition and benefit in form wizard
The wizard steps threw the same ValueError about unsaved instances as I
fixed in the test suite a commit or two ago.
There's a reference to a strange Django behaviour. What I think happened
is that only the workaround forced saving the instances in the database.
It looks like the benefit and condition are only saved in the last step,
so it seems safe to just not populate the unsaved offer instance with
them until we save everything.