Fix depth for CategoryFactory
Treebeard gives a depth of 1 to a root category, at least for the
MP_Node class we're using. As our factory tries to imitate treebeard, we
should do the same.

This was spotted because it caused a bug in some category munging tests
that I was working on for a client.