Set EmailFields to have max_length=75 to side-step migration problem
When the lowest version of Django we support is 1.8, we can remove these
max_lengths and fall back to the length set in Django (now 254). It's
messy to try and support both Django 1.7 and 1.8 with this change as
'makemigrations' will pick up different things depending on which
version of Django is being used.