fix: prevent ThankYouView from just showing the last order to superusers (#4501)
The checkout ThankYouView has logic to allow superusers to view arbitrary orders
to aide in debugging. However, this logic had a bug which cause superusers to
just always be shows the most recent order, even after having just placed an
order themselves.

This commit changes the ThankYouView logic to still allow custom orders to be
shown to superusers, but only when actually requested (via query params).